requirements.txt
# Solo Builder dependencies
# Core CLI (solo_builder_cli.py) uses stdlib only.

# REST API
flask>=3.1

# Anthropic SDK (direct API calls — no subprocess needed)
anthropic>=0.40

# PDF snapshots (optional — gracefully disabled if missing)
matplotlib>=3.10
numpy>=2.2

# Discord bot (optional — only needed for discord_bot/bot.py)
discord.py>=2.0
